There is filtering, forming table, creating new properties. There is no project view

+++ ProjectWatcher/ProjectWatcher/Controllers/ProjectController.cs
@@ -14,12 +14,23 @@ namespace ProjectWatcher.Controllers
     {
         public ActionResult Index(int projectId)
         {
-            return View();         
+            HttpContextWarker cultureProvider = new HttpContextWarker(HttpContext);
+            string culture = cultureProvider.GetCulture();
+            ViewData["projectName"] = ResourcesHelper.GetText("TheProject1", culture);
+            ViewData["lastUser"] = ResourcesHelper.GetText("Admin", culture);
+            ViewData["lastUpdate"] = ResourcesHelper.GetText("date", culture);
+            ViewData["description"] = ResourcesHelper.GetText("Description", culture);
+            ViewData["propertyInfo"] = ResourcesHelper.GetText("PropertyInfo", culture);
+            ProjectModel model = new ProjectModel();
+            model.Id = projectId;
+            ProjectsReader dal = new ProjectsReader();
+            model.ProjectProperties = ProjectHelper.GetProperties(dal, model.Id);
+            return View(model);
         }
 
         public ActionResult AddProperties(int projectId)
         {
-            HttpContextHelper cultureProvider = new HttpContextHelper(HttpContext);
+            HttpContextWarker cultureProvider = new HttpContextWarker(HttpContext);
             string culture = cultureProvider.GetCulture();
             ViewData["projectProperties"] = ResourcesHelper.GetText("ProjectProperties", culture);
             ViewData["availableProperties"] = ResourcesHelper.GetText("AvailableProperties", culture);
@@ -28,35 +39,58 @@ namespace ProjectWatcher.Controllers
             ViewData["nameTitle"] = ResourcesHelper.GetText("Name", culture);
             ViewData["systemNameTitle"] = ResourcesHelper.GetText("SystemName", culture);
             ViewData["typeTitle"] = ResourcesHelper.GetText("Type", culture);
+            ViewData["isImportantTitle"] = ResourcesHelper.GetText("IsImportant", culture);
             ViewData["removeTitle"] = ResourcesHelper.GetText("Remove", culture);
             ViewData["addTitle"] = ResourcesHelper.GetText("Add", culture);
             ProjectModel model = new ProjectModel();
             model.Id = projectId;
             ProjectsReader dal = new ProjectsReader();
-            model.ProjectProperties = ProjectHelper.GetProperties(dal, model.Id);
+            model.ProjectProperties = ProjectHelper.GetVisibleProperties(dal, model.Id);
+            
+            if (model.ProjectProperties == null)
+            {
+                return RedirectToAction("Index", "Projects");
+            }
             PropertyModel[] existingProperties = ProjectHelper.ExcludeProperties(dal, model.ProjectProperties);
             model.OtherProperties = existingProperties;
             return View(model);
         }
 
+        /// <summary>
+        /// Hides property.
+        /// </summary>
+        /// <param name="projectId"></param>
+        /// <param name="systemName">Systemname of hiding property.</param>
+        /// <returns></returns>
+        [HttpPost]
         public ActionResult DeleteProperty(int projectId, string systemName)
         {
             ProjectsReader dal = new ProjectsReader();
-            DAL.Project actualProject = dal.GetProject(projectId);
-            if (User.IsInRole("administrator") || User.Identity.Name == actualProject.Owner)
+            Project actualProject = dal.GetProject(projectId);
+            if (actualProject != null && (User.IsInRole("administrator") || User.Identity.Name == actualProject.Owner))
             {
                 actualProject.DeleteProperty(systemName);                
             }
             return RedirectToAction("AddProperties", new { projectId = projectId });
         }
 
+        [HttpPost]
         public ActionResult AddExistingProperty(int projectId, string systemName)
         {
             ProjectsReader dal = new ProjectsReader();
-            DAL.Project actualProject = dal.GetProject(projectId);
-            if (User.IsInRole("administrator") || User.Identity.Name == actualProject.Owner)
+            Project actualProject = dal.GetProject(projectId);
+            if (actualProject != null && (User.IsInRole("administrator") || User.Identity.Name == actualProject.Owner))
             {
-                actualProject.AddProperty(systemName);
+                try
+                {
+                    actualProject.AddProperty(systemName, true, true, User.Identity.Name);
+                }
+                catch (ConnectionException e)
+                {
+                    HttpContextWarker context = new HttpContextWarker(HttpContext);
+                    TempData["errorMessage"] = ResourcesHelper.GetText("ConnectionError", context.GetCulture());
+                }
+                
             }
             return RedirectToAction("AddProperties", new { projectId = projectId });
         }
@@ -64,12 +98,13 @@ namespace ProjectWatcher.Controllers
         public ActionResult CreationOfNewProperty(int projectId)
         {
             ViewData["projectId"] = projectId;
-            HttpContextHelper cultureProvider = new HttpContextHelper(HttpContext);
+            HttpContextWarker cultureProvider = new HttpContextWarker(HttpContext);
             string culture = cultureProvider.GetCulture();
             ViewData["nameTitle"] = ResourcesHelper.GetText("Name", culture);
             ViewData["sysNameTitle"] = ResourcesHelper.GetText("SystemName", culture);
             ViewData["typeTitle"] = ResourcesHelper.GetText("Type", culture);
-            ViewData["availableValuesTitle"] = ResourcesHelper.GetText("ValuesInEnumeration", culture);            
+            ViewData["availableValuesTitle"] = ResourcesHelper.GetText("ValuesInEnumeration", culture);
+            ViewData["isImportantTitle"] = ResourcesHelper.GetText("IsImportant", culture);
             PropertyModel model = new PropertyModel();
             model.Name = "";
             model.SystemName = "";
@@ -79,13 +114,13 @@ namespace ProjectWatcher.Controllers
             return View(model); 
         }
 
+        [HttpPost]
         public ActionResult CreationOfNewPropertyClick(PropertyModel model, int projectId)
         {
             if (Request.Form.AllKeys.Contains("accept.x"))
             {
-                HttpContextHelper forCulture = new HttpContextHelper(HttpContext);
-                String culture = forCulture.GetCulture();
-                String result = CreateNewProperty(projectId, model, culture);
+                HttpContextWarker contexter = new HttpContextWarker(HttpContext);
+                String result = CreateNewProperty(projectId, model, model.IsImportant, contexter);
                 if (result == null)
                 {
                     return RedirectToAction("AddProperties", new { projectId = projectId });
@@ -93,7 +128,7 @@ namespace ProjectWatcher.Controllers
                 else
                 {
                     TempData["errorMessage"] = result;
-                    return RedirectToAction("CreationOfNewProperty");
+                    return RedirectToAction("CreationOfNewProperty", new{projectId=projectId});
                 }
             }
             return RedirectToAction("AddProperties", new { projectId = projectId });
@@ -102,13 +137,18 @@ namespace ProjectWatcher.Controllers
 
 
 
-        private string CreateNewProperty(int projectId, PropertyModel model, String culture)
+        private string CreateNewProperty(int projectId, PropertyModel model, bool important, HttpContextWarker contexter)
         {
             ProjectsReader dal = new ProjectsReader();
-            String result = null;
+            Project currentProject = dal.GetProject(projectId);
+            String culture = contexter.GetCulture();
+            if (currentProject == null || currentProject.Owner != contexter.User.Identity.Name && !contexter.User.IsInRole("administrator"))
+            {
+                return ResourcesHelper.GetText("NotEnoughRigts", culture);
+            }
             try
             {
-                dal.CreateNewProperty(model.Name, model.SystemName, model.Type, model.AvailableValues.Split('\n'));
+                dal.CreateNewProperty(model.Name, model.SystemName, model.Type, model.AvailableValuesAsArray);
             }
             catch (BadSystemNameException)
             {
@@ -126,16 +166,36 @@ namespace ProjectWatcher.Controllers
             {
                 return ResourcesHelper.GetText("BadDisplayType", culture); 
             }
-            Project currentProject = dal.GetProject(projectId);
-            if (currentProject.Owner == User.Identity.Name)
+            try
             {
-                currentProject.AddProperty(model.SystemName);
-                return null;
+                currentProject.AddProperty(model.SystemName, true, important, contexter.User.Identity.Name);
             }
-            return ResourcesHelper.GetText("NotEnoughRights", culture);            
+            catch (ConnectionException e)
+            {   
+                return ResourcesHelper.GetText("ConnectionError", contexter.GetCulture());
+            }
+            return null;
+        }
+                        
+        
+
+        public PartialViewResult ChangeImportance(String systemName, int projectId)
+        {
+            ProjectsReader dal = new ProjectsReader();
+            Project project = dal.GetProject(projectId);
+            Value toChange = project.Values.FirstOrDefault(x => x.SystemName == systemName);
+            if (toChange == null)
+            {
+                return PartialView("ChangeImportance", Value.CreateValue(systemName, projectId, true, false, ""));
+            }
+            else
+            {
+                Modifier modifier = new Modifier();
+                toChange.Important = !toChange.Important;
+                modifier.Modify(toChange);
+                return PartialView("ChangeImportance", toChange);
+            }
         }
 
-       
-            
     }
 }
